+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 ## 0.92.0 (Unreleased)
 FEATURES:
 * mongodb: update mongodb version in tests from deprecated `4.2` to actual `6.0`
+* organizationmanager: saml user organization membership management via dedicated resource `yandex_organizationmanager_saml_federation_user_account` is implemented
 
 BUG FIXES:
+* organizationmanager: data source `yandex_organizationmanager_saml_federation_user_account` account creation behaviour deprecation notice added
 
 ## 0.91.0 (May 18, 2023)
 FEATURES: